+++ client/src/components/CartItem.js
@@ -2,6 +2,7 @@ import { useEffect, useState } from "react";
 import { useStoreContext } from "../utils/GlobalState";
 import { REMOVE_FROM_CART, UPDATE_CART_QUANTITY } from "../utils/actions";
 import { idbPromise } from "../utils/helpers";
+import OrderSummary from "./OrderSummary";
 
 import {
   Box,
@@ -29,7 +30,6 @@ const CartItem = ({ item }) => {
     });
     console.log("item deleted");
     idbPromise("cart", "delete", { ...item });
-    handleUpdateSummary(item._id, item.count, item.price);
   };
 
 //   const onChange = (e) => {
@@ -52,88 +52,6 @@ const CartItem = ({ item }) => {
 //       idbPromise("cart", "put", { ...item, purchaseQuantity: parseInt(value) });
 //     }
 //   };
-  let initialCartCount =
-    sessionStorage.length > 1 ? sessionStorage.length - 1 : 1;
-  let initialSubTotal = initialCartCount * 2.99;
-  let initialTax = initialSubTotal * 0.0475;
-  let initialTotal = initialSubTotal + initialTax;
-
-  const [itmCount, setCount] = useState([
-    { itmId: 0, Count: state.length, Price: 2.99 * state.length },
-  ]);
-  const [totalCount, setTotalCount] = useState(initialCartCount);
-  const [subTotal, setSubTotal] = useState(initialSubTotal);
-  const [shipAndTax, setTax] = useState(initialTax);
-  const [totalCost, setTotalCost] = useState(initialTotal);
-
-  useEffect(() => {
-    setTotalItemCount();
-  }, [itmCount]);
-  useEffect(() => {
-    set_Sub_Total();
-  }, [totalCount]);
-  useEffect(() => {
-    set_Tax();
-  }, [subTotal]);
-  useEffect(() => {
-    set_Total_Cost();
-  }, [shipAndTax]);
-
-  const setTotalItemCount = () => {
-    let total = 0;
-    for (let i = 0; i < itmCount.length; i++) {
-      total += Number(itmCount[i].Count);
-    }
-    if (total > 0) setTotalCount(total);
-    else setTotalCount(initialCartCount);
-  };
-
-  const set_Sub_Total = () => {
-    let total = 0;
-    for (let i = 0; i < itmCount.length; i++) {
-      total += Number(itmCount[i].Price);
-    }
-    if (total > 0) setSubTotal(total);
-    else setSubTotal(initialSubTotal);
-  };
-
-  const set_Tax = () => {
-    let total = 0;
-    total = total + subTotal * 0.0475;
-    if (total > 0) setTax(total.toFixed(2));
-    else setTax(initialTax);
-  };
-
-  const set_Total_Cost = () => {
-    let total = 0.0;
-    total = parseFloat(subTotal) + parseFloat(shipAndTax);
-    if (total > 0) setTotalCost(total.toFixed(2));
-    else setTotalCost(initialTotal);
-  };
-
-  const handleUpdateSummary = (id, count, price) => {
-    console.log(id)
-    if(itmCount && itmCount.find(x => x.itmId === id)) {
-        const index = itmCount.map(e => e.itmId).indexOf(id)
-        const newCount = [...itmCount]
-        newCount[index].Count = count
-        newCount[index].Price = price
-
-        setCount(newCount)
-    }
-    else {
-
-      const newCount = {
-        itmId: id,
-        Count: count,
-        Price: price
-      }
-
-      const newCnt = [...itmCount, newCount ]
-
-      setCount(newCnt);
-    }
-  }
 
   return (
     <>
@@ -224,44 +142,6 @@ const CartItem = ({ item }) => {
         </Flex>
       </Box>
 
-      <Box mt="50px">
-        <Stack
-          spacing="8"
-          borderWidth="1px"
-          rounded="lg"
-          padding="8"
-          width="300px"
-        >
-          <Heading size="md">Order Summary</Heading>
-          <Stack spacing="6">
-            <InputGroup>
-              <InputLeftAddon children="ItemCount" />
-              <Input placeholder={totalCount} />
-            </InputGroup>
-            <InputGroup>
-              <InputLeftAddon children="Subtotal" />
-              <Input placeholder={subTotal} />
-            </InputGroup>
-            <InputGroup>
-              <InputLeftAddon children="Shipping + Tax" />
-              <Input placeholder={shipAndTax} />
-            </InputGroup>
-            <InputGroup>
-              <InputLeftAddon children="Total" />
-              <Input placeholder={totalCost} />
-            </InputGroup>
-            <Button
-              bg="red"
-              color="white"
-              size="lg"
-              fontSize="md"
-              _hover={{ bg: "brick_red" }}
-            >
-              Checkout
-            </Button>
-          </Stack>
-        </Stack>
-      </Box>
     </>
   );
 };
