+++ client/src/utils/GlobalState.js
@@ -13,7 +13,7 @@ const StoreProvider = ({ value = [], ...props }) => {
     currentCategory: "",
   });
   // use this to confirm it works!
-  console.log(state);
+//   console.log(state);
   return <Provider value={[state, dispatch]} {...props} />;
 };
